src/server/pyproject.toml
[tool.poetry]
name = "bl-ecom-sv"
version = "0.1.0"
description = ""
authors = ["nhqb1010 <quocbao102000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.4"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
